+++ crates/fj-kernel/src/algorithms/intersect/ray_edge.rs
@@ -4,8 +4,8 @@ use fj_math::Segment;
 
 use crate::{
     algorithms::intersect::{HorizontalRayToTheRight, Intersect},
+    geometry::path::SurfacePath,
     objects::HalfEdge,
-    path::SurfacePath,
     storage::Handle,
 };
