+++ crates/fj-kernel/src/geometry/mod.rs
@@ -1,0 +1,3 @@
+//! Types that are tied to objects, but aren't objects themselves
+
+pub mod path;
