+++ crates/fj-kernel/src/geometry/path.rs
