+++ crates/fj-kernel/src/lib.rs
@@ -89,10 +89,10 @@
 
 pub mod algorithms;
 pub mod builder;
+pub mod geometry;
 pub mod insert;
 pub mod iter;
 pub mod objects;
 pub mod partial;
-pub mod path;
 pub mod storage;
 pub mod validate;
